refactor(structure): extract shared WeekDayRow and WorkTimeField to remove duplication
